+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 Changes since 0.6.0 onwards.
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
+0.8.5 (2020-09-02)
+~~~~~~~~~~~~~~~~~~
+* Added funcs.tperi2tzero() and funcs.eclipse_phase()
+* Added "Bayes factors" section to dataset.lmfit_report()
+* Added multivisit.fit_eblm
+* Added pycheops/examples/Notebooks/KELT-11b-tutorial.ipynb
+
 0.8.4 (2020-08-30)
 ~~~~~~~~~~~~~~~~~~
 * Fix parameter hint prefix problem in models (#141)